moved strategyCards into GameComponents
