+++ client/src/KnowledgeBase/StrategyFront.js
@@ -2,7 +2,7 @@ import clsx from 'clsx'
 import { makeStyles } from '@material-ui/core/styles'
 
 import { useTranslation } from '../i18n'
-import { images } from '../gameInfo/strategyCards'
+import { images } from '../GameComponents/gameInfo/strategyCards'
 
 const referenceWidth = 805
 const referenceHeight = 1000
